+++ pkg/config/file.go
@@ -1,7 +1,8 @@
-package io
+package config
+
+import "time"
 
 // Get Date Time prefix for a file
 func GetDateTimePrefixForFile() string {
-	return "temp"
-	// return time.Now().Format("20060102-150405")
+	return time.Now().Format("20060102-150405")
 }
